rename to yield in init
