+++ crates/ruff/resources/test/fixtures/pylint/yield_in_init.py
