+++ crates/ruff/src/registry.rs
@@ -125,7 +125,7 @@ ruff_macros::define_rule_mapping!(
     F842 => rules::pyflakes::rules::UnusedAnnotation,
     F901 => rules::pyflakes::rules::RaiseNotImplemented,
     // pylint
-    PLE0100 => rules::pylint::rules::InitIsGenerator,
+    PLE0100 => rules::pylint::rules::YieldInInit,
     PLE0604 => rules::pylint::rules::InvalidAllObject,
     PLE0605 => rules::pylint::rules::InvalidAllFormat,
     PLE1307 => rules::pylint::rules::BadStringFormatType,
