+++ crates/ruff/src/rules/pylint/rules/mod.rs
@@ -5,7 +5,6 @@ pub use bidirectional_unicode::{bidirectional_unicode, BidirectionalUnicode};
 pub use comparison_of_constant::{comparison_of_constant, ComparisonOfConstant};
 pub use consider_using_sys_exit::{consider_using_sys_exit, ConsiderUsingSysExit};
 pub use global_variable_not_assigned::GlobalVariableNotAssigned;
-pub use init_is_generator::{init_is_generator, InitIsGenerator};
 pub use invalid_all_format::{invalid_all_format, InvalidAllFormat};
 pub use invalid_all_object::{invalid_all_object, InvalidAllObject};
 pub use magic_value_comparison::{magic_value_comparison, MagicValueComparison};
@@ -25,6 +24,7 @@ pub use used_prior_global_declaration::{
 };
 pub use useless_else_on_loop::{useless_else_on_loop, UselessElseOnLoop};
 pub use useless_import_alias::{useless_import_alias, UselessImportAlias};
+pub use yield_in_init::{yield_in_init, YieldInInit};
 
 mod await_outside_async;
 mod bad_str_strip_call;
@@ -33,7 +33,6 @@ mod bidirectional_unicode;
 mod comparison_of_constant;
 mod consider_using_sys_exit;
 mod global_variable_not_assigned;
-mod init_is_generator;
 mod invalid_all_format;
 mod invalid_all_object;
 mod magic_value_comparison;
@@ -49,3 +48,4 @@ mod use_from_import;
 mod used_prior_global_declaration;
 mod useless_else_on_loop;
 mod useless_import_alias;
+mod yield_in_init;
